ReadFromSource/requirements.txt
kafka-python==2.0.3
confluent-kafka==2.8.0
python-dotenv==1.0.1
websockets==15.0
fastapi==0.115.9
uvicorn==0.34.0